pyproject.toml
[project]
name = "imagegen-api"
version = "1.0.0"
description = "ImageGen API Backend - Django Ninja"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1",
    "django-ninja>=1.3",
    "django-cors-headers>=4.6",
    "psycopg[binary]>=3.2",
    "PyJWT>=2.9",
    "bcrypt>=4.2",
    "python-dotenv>=1.0",
    "httpx>=0.28",
    "uvicorn>=0.32",
    "gunicorn>=23.0",
    "cryptography>=44.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.6",
    "pillow>=11.0",
    "python-multipart>=0.0.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-asyncio>=0.24",
    "black>=24.10",
    "ruff>=0.8",
    "mypy>=1.13",
    "django-stubs>=5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
plugins = ["mypy_django_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "core.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py", "*_test.py"]
